+++ rings/src/num_theory/berlekamp_zassenhaus.rs
@@ -53,9 +53,6 @@ some improvements
  - knapsack improvement https://www.math.fsu.edu/~hoeij/knapsack/paper/knapsack.pdf
    This might be of use when I do the LLL bit, especially 10 and 11: http://people.csail.mit.edu/madhu/FT98/
  - brute search optimizations https://www.shoup.net/papers/asz.pdf
- - berlekamp_zassenhaus_algorithm
- - don't use factor by find factor, just do it all in one go and partition the modular factors into true factors
-
 */
 
 use crate::num_theory::modulo::montgomery::MontgomeryModuloOddPrimeStructure;
@@ -1394,6 +1391,7 @@ pub fn factorize_by_berlekamp_zassenhaus_algorithm_naive(
 #[cfg(test)]
 mod van_hoeij_tests {
     use super::*;
+    use crate::num_theory::zimmermann_polys::{p1, p2, p3, p4, p5, p6, p7, p8};
 
     fn poly(coeffs: &[i32]) -> Polynomial<Integer> {
         Polynomial::from_coeffs(coeffs.iter().map(|c| Integer::from(*c)).collect())
@@ -1655,4 +1653,62 @@ mod van_hoeij_tests {
         assert_eq!(factors.len(), 1);
         assert!(Polynomial::<Integer>::structure().equal(&factors[0], &f));
     }
+
+    fn assert_factor_degrees(poly: Polynomial<Integer>, expected_degrees: &[usize]) {
+        let polynomial_ring = Polynomial::<Integer>::structure();
+        let factorization = polynomial_ring.factor(&poly);
+        assert!(polynomial_ring.equal(
+            &poly,
+            &polynomial_ring.factorizations().expand(&factorization)
+        ));
+
+        let mut actual_degrees = factorization
+            .powers()
+            .unwrap()
+            .iter()
+            .flat_map(|(factor, exponent)| {
+                let degree = factor.degree().unwrap();
+                let multiplicity: usize = exponent.try_into().unwrap();
+                std::iter::repeat_n(degree, multiplicity)
+            })
+            .collect::<Vec<_>>();
+        actual_degrees.sort_unstable();
+        assert_eq!(actual_degrees, expected_degrees);
+    }
+
+    macro_rules! zimmermann_factor_test {
+        ($name:ident, $poly:ident, [$($degree:expr),* $(,)?]) => {
+            #[test]
+            #[ignore = "long-running integer polynomial factorization benchmark"]
+            fn $name() {
+                assert_factor_degrees($poly(), &[$($degree),*]);
+            }
+        };
+    }
+
+    zimmermann_factor_test!(
+        factor_zimmermann_p1,
+        p1,
+        [
+            2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 8, 8,
+            8, 8, 8, 8, 8, 8, 8,
+        ]
+    );
+    zimmermann_factor_test!(
+        factor_zimmermann_p2,
+        p2,
+        [2, 2, 12, 12, 12, 12, 24, 24, 24, 24, 24, 24]
+    );
+    zimmermann_factor_test!(
+        factor_zimmermann_p3,
+        p3,
+        [
+            12, 12, 12, 12, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24, 24
+        ]
+    );
+    zimmermann_factor_test!(factor_zimmermann_p4, p4, [66, 396]);
+    zimmermann_factor_test!(factor_zimmermann_p5, p5, [64]);
+    zimmermann_factor_test!(factor_zimmermann_p6, p6, [12, 12, 12, 12, 48, 48]);
+    zimmermann_factor_test!(factor_zimmermann_p7, p7, [384]);
+    zimmermann_factor_test!(factor_zimmermann_p8, p8, [972]);
 }
